:recycle: Rename bookmarkV2 to autoSuggest and move to send folder
