+++ src/components/sendV2/autoSuggest/autoSuggest.css
@@ -1,4 +1,4 @@
-@import '../app/variablesV2.css';
+@import '../../app/variablesV2.css';
 
 .recipientField {
   align-items: center;
