+++ src/components/sendV2/autoSuggest/index.js
@@ -1,14 +1,14 @@
 import React, { Fragment } from 'react';
-import AccountVisual from '../accountVisual/index';
-import { InputV2 } from '../toolbox/inputsV2';
-import keyCodes from '../../constants/keyCodes';
-import svg from '../../utils/svgIcons';
-import SpinnerV2 from '../spinnerV2/spinnerV2';
-import Feedback from '../toolbox/feedback/feedback';
-import styles from './bookmark.css';
+import AccountVisual from '../../accountVisual/index';
+import { InputV2 } from '../../toolbox/inputsV2';
+import keyCodes from '../../../constants/keyCodes';
+import svg from '../../../utils/svgIcons';
+import SpinnerV2 from '../../spinnerV2/spinnerV2';
+import Feedback from '../../toolbox/feedback/feedback';
+import styles from './autoSuggest.css';
 
 // eslint-disable-next-line complexity
-class Bookmark extends React.Component {
+class AutoSuggest extends React.Component {
   constructor(props) {
     super(props);
 
@@ -194,4 +194,4 @@ class Bookmark extends React.Component {
   }
 }
 
-export default Bookmark;
+export default AutoSuggest;
